+++ src/main/java/frc/robot/subsystems/ballevator/BallevatorFactory.java
@@ -10,7 +10,8 @@ package frc.robot.subsystems.ballevator;
 import org.slf4j.Logger;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.commands.ballevator.BallevatorSimpleJoystickControl;
+
+import frc.robot.commands.ballevator.BallevatorSimpleManual;
 import frc.robot.subsystems.SubsystemNames;
 import frc.robot.telemetry.TelemetryNames;
 
@@ -64,7 +65,7 @@ public class BallevatorFactory {
             SmartDashboard.putBoolean(TelemetryNames.Ballevator.status, true);
         }
 
-        ourInstance.setDefaultCommand(new BallevatorSimpleJoystickControl());
+        ourInstance.setDefaultCommand(new BallevatorSimpleManual());
     }
 
     /**
